refactor: extract duplicate Stripe search logic into shared stripe-utils module
Move duplicated `findExistingProduct`, `findExistingPrice`, search query escape
logic, `fetchCurrentResource`, and `normalizeResource` from `deployer.ts` and
`diff.ts` into a new `packages/cli/src/engine/stripe-utils.ts` module. Both
files now import and delegate to the shared utilities.
